Refactor get_defs
