+++ lsp/nls/src/requests/goto.rs
@@ -1,75 +1,9 @@
 use lsp_server::{RequestId, Response, ResponseError};
 use lsp_types::{GotoDefinitionParams, GotoDefinitionResponse, Location, ReferenceParams};
-use nickel_lang_core::{
-    position::RawSpan,
-    term::{record::FieldMetadata, RichTerm, Term, UnaryOp},
-};
+use nickel_lang_core::position::RawSpan;
 use serde_json::Value;
 
-use crate::{
-    cache::CacheExt,
-    diagnostic::LocationCompat,
-    field_walker::{Def, FieldResolver},
-    identifier::LocIdent,
-    server::Server,
-};
-
-fn get_defs(term: &RichTerm, ident: Option<LocIdent>, server: &Server) -> Option<Vec<RawSpan>> {
-    let resolver = FieldResolver::new(server);
-    let ret = match (term.as_ref(), ident) {
-        (Term::Var(id), _) => {
-            let id = LocIdent::from(*id);
-            let def = server.analysis.get_def(&id)?;
-            let cousins = resolver.cousin_defs(def);
-            if cousins.is_empty() {
-                vec![def.ident().pos.unwrap()]
-            } else {
-                cousins
-                    .into_iter()
-                    .filter_map(|(loc, _)| loc.pos.into_opt())
-                    .collect()
-            }
-        }
-        (Term::Op1(UnaryOp::StaticAccess(id), parent), _) => {
-            let parents = resolver.resolve_record(parent);
-            parents
-                .iter()
-                .filter_map(|parent| {
-                    parent
-                        .field_loc(id.ident())
-                        .and_then(|def| def.pos.into_opt())
-                })
-                .collect()
-        }
-        (Term::LetPattern(_, pat, value, _), Some(hovered_id)) => {
-            let (mut path, _, _) = pat
-                .matches
-                .iter()
-                .flat_map(|m| m.to_flattened_bindings())
-                .find(|(_path, bound_id, _)| bound_id.ident() == hovered_id.ident)?;
-            path.reverse();
-            let (last, path) = path.split_last()?;
-            let path: Vec<_> = path.iter().map(|id| id.ident()).collect();
-            let parents = resolver.resolve_path(value, path.iter().copied());
-            parents
-                .iter()
-                .filter_map(|parent| {
-                    parent
-                        .field_loc(last.ident())
-                        .and_then(|def| def.pos.into_opt())
-                })
-                .collect()
-        }
-        (Term::ResolvedImport(file), _) => {
-            let pos = server.cache.terms().get(file)?.term.pos;
-            vec![pos.into_opt()?]
-        }
-        _ => {
-            return None;
-        }
-    };
-    Some(ret)
-}
+use crate::{cache::CacheExt, diagnostic::LocationCompat, server::Server};
 
 fn ids_to_locations(ids: impl IntoIterator<Item = RawSpan>, server: &Server) -> Vec<Location> {
     let mut spans: Vec<_> = ids.into_iter().collect();
@@ -97,7 +31,7 @@ pub fn handle_to_definition(
 
     let locations = server
         .lookup_term_by_position(pos)?
-        .and_then(|term| get_defs(term, ident, server))
+        .map(|term| server.get_defs(term, ident))
         .map(|defs| ids_to_locations(defs, server))
         .unwrap_or_default();
 
@@ -125,39 +59,13 @@ pub fn handle_references(
     // so first find the definitions and then find their usages.
     let term = server.lookup_term_by_position(pos)?;
     let mut def_locs = term
-        .and_then(|term| get_defs(term, ident, server))
+        .map(|term| server.get_defs(term, ident))
         .unwrap_or_default();
 
     // Maybe the position is pointing straight at the definition already.
     // In that case, def_locs won't have the definition yet; so add it.
-    if let Some(id) = server.lookup_ident_by_position(pos)? {
-        if let Some(span) = id.pos.into_opt() {
-            def_locs.push(span);
-            if let Some(parent) = term {
-                // If `id` is a field name in a record, we can search through cousins
-                // to find more definitions.
-                if matches!(parent.as_ref(), Term::RecRecord(..) | Term::Record(_)) {
-                    let def = Def::Field {
-                        ident: id,
-                        value: None,
-                        record: parent.clone(),
-                        metadata: FieldMetadata::default(),
-                    };
-                    let resolver = FieldResolver::new(server);
-                    let cousins = resolver.cousin_defs(&def);
-                    def_locs.extend(
-                        cousins
-                            .into_iter()
-                            .filter_map(|(loc, _)| loc.pos.into_opt()),
-                    )
-                }
-            }
-        }
-    }
+    def_locs.extend(ident.and_then(|id| id.pos.into_opt()));
 
-    // TODO: This usage map is based only on static scoping, and not on our "extended"
-    // scopes that we build up dynamically based on merges. Improving this probably
-    // requires building the extended scopes at static analysis time.
     let mut usages: Vec<_> = def_locs
         .iter()
         .flat_map(|id| server.analysis.get_usages(id))
